Started transforming basic news portal to OKOSL homework revision portal.
For a lab for Advanced databases, I implemented a very simple Mongo-based news portal in Flask.
For OKOSL AY 2019/2020 we're implementing a semi-automatic homework revision process, so this
web portal could be repurposed to serve as a platform for facilitating this revision.

Currently, a data model is added, basic Postgres config is added, but routes.py still
needs to be repurposed to reflect the changes.