Moved user creation from __init__ to db_helper
User creation should be interactive, and not hardcoded, so I moved the functionality
to the script db_helper. This can probably be done better (e.g. using Flask-Script), but I
ran into problems, so I will solve it at a later date.
For migrations to work, the models must be imported somewhere. This was done in __init__.py