Changed default HW name simply to "DZ"
The motivation behind this is that homeworks are uniquely identified by
their year and their ordinal number, so names like "DZ_2019-3" can be
generated on-the-fly. However, it is possible that in the future we will
extend the app to support lab exercises ("LAB") in some way, so the 'name'
attribute remains here.