Moving a solution to a new SG now works.
Rendering the subtask template decomposed so that the SG list is in its own Jinja2 template, and is being rendered
SG by SG through Python. This enabled me to use an AJAX call to add a new SG when a solution is moved to "empty space"
i.e. to a new SG of its own.